Ycbran

+++ src/components/ReactFlow/DataTableNode.tsx
@@ -26,12 +26,21 @@ export default function DataTableNode({ data }) {  //this 'data' is created and 
   let secondaryFirstRow = []
   let RowData = Object.values(tableData[1]);
 
- //Used to grab the primary key value in the Table
-  let schemaName = schemaStore[`public.${tableName}`];
-  let PK = null;
-  for(let key in schemaName){
-    if(schemaName[key]['IsPrimaryKey']) PK = schemaName[key].field_name;
-  }
+ //Used to grab the primary key and foreign keys column in the Table
+ let schemaName = schemaStore[`public.${tableName}`];
+ let PK = null;
+ let FK = null
+ let pkVals = new Set();
+ for(let key in schemaName){
+   if(schemaName[key]['IsForeignKey']) FK = schemaName[key].field_name;
+   if(schemaName[key]['IsPrimaryKey']) PK = schemaName[key].field_name;
+ }
+ 
+//loop through all of RowData, grab each primary key value and store it in object<pkVals>
+for(let i = 0; i < RowData.length; i++){
+  pkVals.add(RowData[i][PK]);
+}
+
 
   if (schemaName !== undefined) {
     secondaryFirstRow = Object.keys(schemaStore['public.' + tableName]);
@@ -53,17 +62,24 @@ export default function DataTableNode({ data }) {  //this 'data' is created and 
   }, [dataStore]);
 
 
-  const deleteRow = async (value,index,id) => {
-    restRowsData = restRowsData.slice(0,index).concat(restRowsData.slice(index+1,restRowsData.length))
-    setDataStore({...dataStore,[id]:restRowsData});
-    const sendDeleteRequest = fetch(`/api/${dbCredentials.db_type}/deleteRow`,{
-      method:'DELETE',
-      headers:{
-        'Content-Type':'application/json'
-      },
-      body:JSON.stringify({tableName : tableName, primaryKey: PK, value: value[PK] })
-    })
+ const deleteRow = async (value,index,id) => {
+ 
+  restRowsData = restRowsData.slice(0,index).concat(restRowsData.slice(index+1,restRowsData.length))
+  if(value[FK]!== null){
+    alert(`Can't Delete Foreign Key: ${FK}`);
+    throw new Error('Duplicate Primary Key');
+  }
+   setDataStore({...dataStore,[id]:restRowsData});
 
+
+  const sendDeleteRequest = fetch(`/api/${dbCredentials.db_type}/deleteRow`,{
+    method:'DELETE',
+    headers:{
+      'Content-Type':'application/json'
+    },
+    body:JSON.stringify({tableName : tableName, primaryKey: PK, value: value[PK] })
+  })
+  
   ////////////////// Fetch path: /api/delete ///////////////////
   // {
   //  tableName: name of table,
@@ -170,6 +186,8 @@ export default function DataTableNode({ data }) {  //this 'data' is created and 
                       id={tableName}
                       index={index}
                       deleteRow={deleteRow}
+                      FK={FK}
+                      PK={[PK,pkVals]}
                     />
                   )} 
                 )}
